+++ CHANGELOG.md
@@ -31,6 +31,10 @@ Dynamic Doors
 - Dynamic doors can teleport to any dimension, or position.
 - Dynamic doors also have onEnter, or beforeEnter callbacks for additional door functionality.
 - Includes better handling of ytyp, and ipl loading; no more falling through map.
+
+Drop Tables
+- Create a drop table and pull random loot from it.
+- Athena.systems.dropTable.get
 ```
 
 ## 5.0.0